pyproject.toml
[project]
name = "icland"
version = "0.1.0a2"
description = "Recreating Google DeepMind's XLand RL environment in JAX"
readme = "README.md"
authors = []
requires-python = ">=3.13"
license_files = ["LICENSE"]
dependencies = [
    "beartype>=0.19.0",
    "brax>=0.12.1",
    "flask>=3.1.0",
    "flask-cors>=5.0.0",
    "flask-socketio>=5.5.1",
    "flax>=0.10.2",
    "gputil>=1.4.0",
    "hilbertcurve>=2",
    "imageio[ffmpeg]>=2.37.0",
    "jax>=0.5.0 ; sys_platform != 'linux'",
    "jax[cuda12]>=0.5.0 ; sys_platform == 'linux'",
    "jaxtyping>=0.2.37",
    "matplotlib>=3.10.0",
    "mujoco>=3.2.6",
    "mujoco-mjx>=3.2.6",
    "numpy>=2.2.2",
    "numpy-stl>=3.2.0",
    "opencv-python>=4.11.0.86",
    "py-cpuinfo>=9.0.0",
    "pygame>=2.6.1",
    "pylatex>=1.4.2",
    "pynvml>=12.0.0",
    "requests>=2.32.3",
    "scipy>=1.15.1",
    "types-setuptools>=75.8.0.20250110",
    "websocket-client>=1.8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "mypy>=1.14.1",
    "pre-commit>=4.0.1",
    "pytest>=8.3.4",
    "pytest-cov>=6.0.0",
    "ruff>=0.9.1",
]
docs = [
    "furo>=2024.8.6",
    "myst-parser>=4.0.1",
    "sphinx>=8.2.1",
    "sphinx-copybutton>=0.5.2",
    "sphinx-design>=0.6.1",
    "sphinxext-opengraph>=0.9.1",
]

[tool.ruff]
target-version = "py313"

[tool.ruff.lint]
select = [
    "D",    # Enable pydocstyle checks
    "F401", # Remove unused imports
    "UP",   # pyupgrade: rewrite old syntax
]
extend-select = ["I"] # Isort imports

[tool.ruff.lint.pydocstyle]
convention = "google"

# "import-untyped" - Mujoco lacks supports for types
# https://github.com/google-deepmind/mujoco/issues/244#issuecomment-1111010344
# "no-any-return" - mypy support for jitted function assumes it returns Any
# Rely on runtime type checking here instead
[tool.mypy]
disable_error_code = ["import-untyped", "no-any-return"]
